pyproject.toml
[project]
name = "DynSeg"
version = "0.1.0"
description = "DynSeg: Dynamic Separable Convolutions for Zero-Shot Referring Image Segmentation."
requires-python = ">=3.10"
dependencies = [
    "abcplus>=0.1.0",
    "accelerate>=1.12.0",
    "albumentations>=2.0.8",
    "datasets>=4.4.1",
    "einops>=0.8.1",
    "evaluate>=0.4.6",
    "filelock>=3.20.0",
    "huggingface-hub>=0.36.0",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "safetensors>=0.7.0",
    "scikit-image>=0.25.2",
    "tokenizers>=0.22.1",
    "toml>=0.10.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wandb>=0.23.1",
    "webdataset>=1.0.2",
]
